# tests/atlas_contract/openapi.yaml
openapi: 3.1.0
info:
  title: Energy Atlas Collector API
  version: 1.0.0-e0
  summary: Ingest and read API for Energy Atlas sites (E0).
  description: |
    Single source of truth for the collector (Go) and the edge agent `atlas_sink` (Python).
    Architecture: docs/ARCHITECTURE.md. Decisions: docs/adr/.

    Conventions
    - All timestamps are RFC 3339 in UTC with `Z` (invariant 1). Non-UTC offsets are rejected (422).
    - Aggregation window is fixed at 300 s. `window_start` must be a multiple of 300 s since the
      Unix epoch (a full 5 minutes UTC).
    - Errors are RFC 9457 `application/problem+json` (`Problem`). `type` is
      `https://api.fleet.datamindz.io/problems/<slug>`; slugs are listed in `Problem.type`. The
      base is an opaque identifier kept stable across the host change (ADR-0016 §3), not the API host.
    - Every response to a site-signed POST carries `Config-ETag` (effective config ETag).
    - Request bodies may be `Content-Encoding: gzip`. Limit: 1 MiB after decompression (413).
    - Writes are idempotent on natural keys; a repeated request never changes stored rows.
    - Unknown fields are rejected (`additionalProperties: false`) with 422 naming the field.
    - Malformed or missing path/query/header parameters (e.g. non-UUID `site_id`, missing `from`)
      → 422 `invalid-field`, `errors[].field` = `path.<name>` / `query.<name>` / `header.<Name>`
      (never 400, never 404). Every operation with parameters lists 422.
...

...
  contact:
...
servers:
...
security: []
paths:
...
  /v1/sites:
    post:
      tags: [site]
      operationId: registerSite
      summary: Register a site public key under an enrollment credential (S-001)
      description: |
        `source` is taken from the credential, never from the body (a `source` field is an unknown
        field → 422). Same `public_key` again → 200 with the existing site if the credential has the
        same source, else 409 `conflict`, site unchanged (`source-mismatch` is only the 403 of
        `siteSignature`, ADR-0003). A `public_key` whose site key was revoked
        (`revokeSiteKey`) → 409 `site-key-revoked`, nothing created or changed; checked after auth
        and before the source check; the edge must generate a new key to re-enrol (ADR-0011).
        Rate limit 10 requests/hour per client IP (counted before auth) → 429.
      security: [{ enrollmentBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteRegistration' }
      responses:
        '201':
          description: New site created.
          headers:
            Config-ETag: { $ref: '#/components/headers/ConfigETag' }
            Location:
              description: /v1/sites/{site_id}
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SiteRegistered' }
        '200':
          description: Key already registered; existing site returned, nothing created.
          headers:
            Config-ETag: { $ref: '#/components/headers/ConfigETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SiteRegistered' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
    get:
      tags: [operator]
      operationId: listSites
      summary: List sites with source, last-seen and alert state (S-010)
      security: [{ operatorBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/SourceFilter'
      responses:
        '200':
          description: Sites ordered by registered_at ascending.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [sites]
                properties:
                  sites:
                    type: array
                    items: { $ref: '#/components/schemas/SiteSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
components:
  securitySchemes:
    siteSignature:
      type: apiKey
      in: header
      name: Signature
      description: |
        RFC 9421 signature with the site Ed25519 key; requires `Signature-Input` and, for requests
        with a body, `Content-Digest` (RFC 9530, sha-256). Profile: ADR-0003.
    operatorBearer:
      type: http
      scheme: bearer
      description: Operator admin token (env ATLAS_OPERATOR_TOKEN).
    enrollmentBearer:
      type: http
      scheme: bearer
      description: Enrollment credential secret `eac_<credential_id>_<secret>`.
  parameters:
...
    SourceFilter:
      name: source
      in: query
      required: false
      description: Only sites with this source. Absent = all.
      schema: { $ref: '#/components/schemas/Source' }
...
  headers:
    ConfigETag:
      description: ETag of the site's effective config (quoted sha256 hex of canonical JSON). Differs from the local one → GET config.
      schema: { type: string, examples: ['"3b1f…"'] }
...
    RetryAfter:
      description: Seconds until retry is allowed.
      schema: { type: integer, minimum: 1 }
  responses:
    BadRequest:
      description: Malformed JSON or body not decodable (`invalid-body`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid credential/signature (`unauthenticated`, `credential-invalid`, `signature-invalid`, `signature-expired`, `digest-mismatch`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not allowed (`site-mismatch`, `source-mismatch`, `operator-only`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Conflicts with stored state (`conflict`; `site-key-revoked` on `registerSite` only).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooLarge:
      description: Body > 1 MiB decompressed, or batch > 288 items (`payload-too-large`). Nothing stored.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Unprocessable:
      description: |
        Validation failed (`invalid-field`); `errors` names every offending body field or
        path/query/header parameter (e.g. a `site_id` that is not a UUID → `path.site_id`). Nothing stored.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Rate limit exceeded (`rate-limited`).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    SiteId:
      type: string
      format: uuid
      description: UUIDv7 assigned by the collector.
    Source:
      type: string
      enum: [real, simulated]
      description: Provenance; fixed by the enrollment credential at registration, immutable.
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339, UTC, `Z` suffix.
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            `https://api.fleet.datamindz.io/problems/<slug>`, slug one of: invalid-body,
            invalid-field, payload-too-large, unsupported-encoding, unauthenticated,
            credential-invalid, signature-invalid, signature-expired, digest-mismatch,
            site-mismatch, source-mismatch, site-key-revoked, operator-only, not-found, method-not-allowed,
            conflict, rate-limited, unavailable, internal.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Present for `invalid-field` (and `payload-too-large` when a batch limit is hit).
          items: { $ref: '#/components/schemas/FieldError' }
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
          description: |
            Body field: JSON Pointer (RFC 6901), e.g. `/windows/3/window_start`.
            Parameter: `<in>.<name>` with the name as declared in this contract —
            `path.<name>` (e.g. `path.site_id`, `path.credential_id`), `query.<name>`
            (e.g. `query.from`), `header.<Name>` (canonical case, e.g. `header.If-None-Match`).
            Malformed or missing parameters use reason `wrong_type` or `required`.
        reason:
          type: string
          description: |
            Machine-readable code: required, wrong_type, unknown_field, out_of_range,
            not_aligned_5min, not_aligned_hour, not_utc, in_future, duplicate_key_in_batch,
            too_many_items, not_multiple_of_300, invalid_h3, forbidden_identifying_data,
            invalid_enum, from_not_before_to, range_too_large, unknown_metric, invalid_tz,
            invalid_public_key.
        detail: { type: string }
...
    SiteRegistration:
      type: object
      additionalProperties: false
      required: [public_key]
      properties:
        public_key:
          type: string
          description: Ed25519 public key, 32 bytes, base64url without padding.
          pattern: '^[A-Za-z0-9_-]{43}$'
    SiteRegistered:
      type: object
      additionalProperties: false
      required: [site_id, source, registered_at]
      properties:
        site_id: { $ref: '#/components/schemas/SiteId' }
        source: { $ref: '#/components/schemas/Source' }
        registered_at: { $ref: '#/components/schemas/Timestamp' }
    Freshness:
      type: object
      additionalProperties: false
      required: [last_seen, last_window_start, last_received_at]
      description: All null when the site has no telemetry window yet. Hourly backfill rows do not count.
      properties:
        last_seen:
          description: End of the latest stored window (last_window_start + 300 s).
          oneOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }]
        last_window_start:
          oneOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }]
        last_received_at:
          description: Server receive time of the latest window (by window_start).
          oneOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }]
...
    SiteSummary:
      type: object
      additionalProperties: false
      required: [site_id, source, registered_at, key_revoked, freshness, alert_active]
      properties:
        site_id: { $ref: '#/components/schemas/SiteId' }
        source: { $ref: '#/components/schemas/Source' }
        registered_at: { $ref: '#/components/schemas/Timestamp' }
        key_revoked: { type: boolean }
        freshness: { $ref: '#/components/schemas/Freshness' }
        alert_active: { type: boolean }
...